Fixed methods

+++ src/main/java/fadesp/paymentrestapi/service/PaymentService.java
@@ -1,9 +1,6 @@
 package fadesp.paymentrestapi.service;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.Objects;
-import java.util.Optional;
+import java.util.*;
 
 import fadesp.paymentrestapi.model.Payment;
 import fadesp.paymentrestapi.repository.PaymentRepository;
@@ -16,8 +13,24 @@ public class PaymentService {
     @Autowired
     private PaymentRepository repository;
 
-    public List<Payment> savePayment(List<Payment> payments){
-        return repository.saveAll(payments);
+    public List<Payment> savePayment(List<Payment> payments) {
+        List<Payment> savedPayments = new ArrayList<>();
+        List<String> validPaymentMethods = Arrays.asList("boleto", "pix", "cartao_credito", "cartao_debito");
+
+        payments.forEach(payment -> {
+            if(validPaymentMethods.equals(payment.getPaymentMethod()) || payment.getPaymentMethod() == null || payment.getPaymentMethod() == "" ){
+                System.out.println("Verifique seus registro, pois um metodo de pagamento informando não existe na base dados ou esta vasil");
+                return;
+            }else if (payment.getPaymentMethod() != null && payment.getPaymentMethod().equals("cartao_credito")
+                        || payment.getPaymentMethod().equals("cartao_debito")) {
+                    if (payment.getCardNumber() == null || payment.getCardNumber().isEmpty()){
+                        System.out.println("Verifique seus registro, pois ou o metodo de pagamento ou numero de cartão estão preenchido incorretamente");
+                        return;
+                    }
+                }
+            savedPayments.add(repository.save(payment));
+        });
+        return savedPayments;
     }
 
     public List<Payment> getPayments(){
@@ -61,7 +74,20 @@ public class PaymentService {
         return repository.save(payment);
     }
 
-    public void deletePayment(Integer debitCode, String paymentStatus){
-            repository.deleteById(debitCode);
+    public void deletePayment(Integer debitCode){
+        Optional<Payment> paymentOptional = repository.findById(debitCode);
+        if (paymentOptional.isPresent()) { // verifica se o pagamento existe
+            Payment payment = paymentOptional.get();
+            if (payment.getPaymentStatus().equals("pendente_processamento")) { // verifica se o status é pendente de processamento
+                repository.deleteById(debitCode);
+                System.out.println("Pagamento foi "+debitCode+" excluido com sucesso");
+            } else {
+                System.out.println("Não é possível excluir um pagamento com status diferente de 'pendente_processamento'");
+                return;
+            }
+        } else {
+            System.out.println("Pagamento não encontrado");
+            return;
+        }
     }
 }
